backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssbeautifier==1.15.1
defusedxml==0.7.1
Django==5.1.1
django-allauth==0.61.1
django-cors-headers==4.4.0
django-rename-app==0.1.7
django-rest-framework==0.1.0
django-sass==1.1.0
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djlint==1.35.2
drf-spectacular==0.27.2
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.10
inflection==0.5.1
jsbeautifier==1.15.1
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
libsass==0.23.0
oauthlib==3.2.2
pathspec==0.12.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-magic==0.4.27
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
stripe==11.1.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
